feat: add guards for demo account

+++ actions/safe-delete-account.ts
@@ -1,5 +1,6 @@
 "use server";
 
+import { isAllowed } from "@/data-access/permission";
 import { UserTable } from "@/data-access/user";
 import { ERROR_MESSAGES } from "@/lib/error-messages";
 import { getSession } from "@/lib/get-session";
@@ -10,6 +11,7 @@ export const deleteAccount = async () => {
   // 1 - Get session data
   const session = await getSession();
   if (!session) return ServerResponse.fail(ERROR_MESSAGES.UNAUTHORIZED);
+  if (!isAllowed(session.user.id)) return ServerResponse.fail(ERROR_MESSAGES.UNAUTHORIZED);
 
   // 2 - Send request to DAL
   const response = await UserTable.deleteAccount({
